refactor(plugin-swarm): rename lib files and corresponding tests
